requirements.txt
streamlit==1.39.0
tensorflow==2.13.0
numpy==1.24.3
pillow==10.4.0
